api/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
redis>=4.5.0
python-dotenv==1.0.0
